Adding support for default in ES Modules
When adding views it will check for the default property
